new Commit
Administrator and staff are a distributed application, developed to maintain the details of employees working in any organization. It maintains the information about the personal details of their employees, also the details about the payroll system which enable to generate the pays lip. The application is actually a suite of applications developed using Java.